+++ Apps/Contoso.WinForms.Demo.DotNetCore/Settings.Designer.cs
@@ -46,5 +46,17 @@ namespace Contoso.WinForms.Demo.DotNetCore {
                 this["TextErrorAttachments"] = value;
             }
         }
+        
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("0")]
+        public long StorageMaxSize {
+            get {
+                return ((long)(this["StorageMaxSize"]));
+            }
+            set {
+                this["StorageMaxSize"] = value;
+            }
+        }
     }
 }