+++ Apps/Contoso.WinForms.Demo/MainForm.cs
@@ -28,11 +28,16 @@ namespace Contoso.WinForms.Demo
             textAttachments = Settings.Default.TextErrorAttachments;
             TextAttachmentTextBox.Text = textAttachments;
             FileAttachmentPathLabel.Text = fileAttachments;
+            if (Settings.Default.StorageMaxSize > 0)
+            {
+                StorageMaxSizeTextBox.Text = Settings.Default.StorageMaxSize.ToString();
+            }
         }
 
         private void UpdateState()
         {
             AppCenterEnabled.Checked = AppCenter.IsEnabledAsync().Result;
+            AppCenterAllowNetworkRequest.Checked = AppCenter.IsNetworkRequestsAllowed;
             AnalyticsEnabled.Checked = Analytics.IsEnabledAsync().Result;
             CrashesEnabled.Checked = Crashes.IsEnabledAsync().Result;
             AnalyticsEnabled.Enabled = AppCenterEnabled.Checked;
@@ -42,6 +47,11 @@ namespace Contoso.WinForms.Demo
         private void AppCenterEnabled_CheckedChanged(object sender, EventArgs e)
         {
             AppCenter.SetEnabledAsync(AppCenterEnabled.Checked).Wait();
+        }
+
+        private void AppCenterAllowNetworkRequest_CheckedChanged(object sender, EventArgs e)
+        {
+            AppCenter.IsNetworkRequestsAllowed = AppCenterAllowNetworkRequest.Checked;
         }
 
         private void AnalyticsEnabled_CheckedChanged(object sender, EventArgs e)
@@ -210,6 +220,16 @@ namespace Contoso.WinForms.Demo
             {
                 TrackException(e);
             }
+        }
+
+        private void SaveStorageSize_Click(object sender, EventArgs e)
+        {
+            var storageSize = StorageMaxSizeTextBox.Text;
+            var size = 10L * 1024 * 1024;
+            long.TryParse(storageSize, out size);
+            AppCenter.SetMaxStorageSizeAsync(size);
+            Settings.Default.StorageMaxSize = size;
+            Settings.Default.Save();
         }
     }
 }