+++ SDK/AppCenter/Microsoft.AppCenter.Windows.Shared/Ingestion/Http/IngestionHttp.cs
@@ -31,6 +31,11 @@ namespace Microsoft.AppCenter.Ingestion.Http
         public IServiceCall Call(string appSecret, Guid installId, IList<Log> logs)
         {
             var call = new ServiceCall(appSecret, installId, logs);
+            if (!IsEnabled)
+            {
+                call.SetException(new NetworkIngestionException(new Exception("SDK is in offline mode.")));
+                return call;
+            }
             CallAsync(appSecret, installId, logs, call.CancellationToken).ContinueWith(task =>
             {
                 // Cancellation token is already shared.
@@ -116,6 +121,11 @@ namespace Microsoft.AppCenter.Ingestion.Http
         public void Dispose()
         {
             _httpNetwork.Dispose();
+        }
+
+        public bool IsEnabled 
+        { 
+            get => AppCenter.PlatformIsNetworkRequestsAllowed; 
         }
     }
 }